src/requirement-token-server.txt
annotated-types==0.5.0
anyio==3.7.1
fastapi==0.100.0
grpcio==1.56.0
grpcio-tools==1.56.0
idna==3.4
protobuf==4.23.4
pydantic==2.0.2
pydantic-settings==2.0.1
pydantic_core==2.1.2
python-dotenv==1.0.0
redis==4.6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
